Fix 'p4sync' auto complete and validate checks.
